requirements.txt
absl-py==0.15.0
aiohttp==3.8.4
aiosignal==1.3.1
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
azure-cognitiveservices-speech==1.22.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.1
bleach==3.2.1
blobfile==2.0.1
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.4
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
cmake==3.21.3
colorama==0.4.4
cycler==0.10.0
dataclasses-json==0.5.7
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.2
dlib==19.22.99
docopt==0.6.2
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==2.2.2
flatbuffers==1.12
frozenlist==1.3.3
gast==0.4.0
google-api-core==1.34.0
google-api-python-client==2.57.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-cloud==0.34.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
google-pasta==0.2.0
googleapis-common-protos==1.56.4
gpt-index==0.4.21
greenlet==2.0.2
grpcio==1.51.1
grpcio-status==1.48.2
gspread==5.7.2
h5py==3.1.0
HTMLParser==0.0.2
httplib2==0.20.4
huggingface-hub==0.12.1
idna==3.3
imageio==2.9.0
importlib-metadata==4.6.1
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.1.0
Js2Py==0.71
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langchain==0.0.101
lxml==4.9.2
Markdown==3.3.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.4.2
mistune==0.8.4
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.6.3
nltk==3.8.1
notebook==6.1.6
numpy==1.24.2
oauthlib==3.1.1
openai==0.27.0
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==21.0
pandas==1.5.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.3.1
pipwin==0.5.2
platformdirs==2.0.2
plotly==5.14.0
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.20.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycparser==2.20
pycryptodomex==3.17
pydantic==1.10.5
Pygments==2.7.3
pyjsparser==2.7.1
pyngrok==5.2.1
pyparsing==2.4.7
PyPrind==2.11.3
pyrsistent==0.17.3
pySmartDL==1.3.4
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==6.0
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2022.10.31
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
tenacity==8.1.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==3.0.0
tifffile==2021.10.12
tiktoken==0.3.0
tokenizers==0.13.2
toml==0.10.2
torch==1.10.0
tornado==6.1
tqdm==4.65.0
traitlets==5.0.5
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.5.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.5.1
wit==6.0.1
wrapt==1.12.1
yarl==1.8.2
zipp==3.5.0
